# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Norv_ai – Smart IT Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Profile / Onboarding ──────────────────────────────────────────────────
...
  /curricula:
    get:
      operationId: getCurricula
      tags: [curricula]
      summary: List all curricula
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Curriculum"

    post:
      operationId: createCurriculum
      tags: [curricula]
      summary: Create curriculum (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurriculumInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Curriculum"
  /curricula/{id}:
    patch:
      operationId: updateCurriculum
      tags: [curricula]
      summary: Update curriculum (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurriculumUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Curriculum"

    delete:
      operationId: deleteCurriculum
      tags: [curricula]
      summary: Delete curriculum (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ── AI Chatbot ────────────────────────────────────────────────────────────
  /chat/sessions:
    get:
      operationId: getChatSessions
      tags: [chat]
      summary: List chat sessions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"

    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{id}:
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /chat/sessions/{id}/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: List messages in a session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message and get AI reply
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ── Smart Summary ─────────────────────────────────────────────────────────
...
  /debug/analyze:
    post:
      operationId: analyzeCode
      tags: [debug]
      summary: Analyze code with AI, detect errors, score efficiency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebugInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSession"
...
  /debug/sessions/{id}:
    get:
      operationId: getDebugSession
      tags: [debug]
      summary: Get a single debug session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugSession"

  # ── Quiz Generator ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [username, password, fullName]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        fullName:
          type: string
          minLength: 1
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, fullName, role, setupComplete, createdAt]
      properties:
        id:
          type: number
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [student, admin]
        setupComplete:
          type: boolean
        university:
          type: ["string", "null"]
        major:
          type: ["string", "null"]
        yearOfStudy:
          type: ["number", "null"]
        specialization:
          type: ["string", "null"]
        skillLevel:
          type: ["string", "null"]
        knownLanguages:
          type: array
          items:
            type: string
        subscriptionActive:
          type: boolean
        subscriptionExpiry:
          type: ["string", "null"]
        createdAt:
          type: string

    # Profile / Onboarding
...
    Curriculum:
      type: object
      required: [id, name, specialization, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        specialization:
          type: string
        lectureCount:
          type: number
        createdAt:
          type: string
    CurriculumInput:
      type: object
      required: [name, specialization]
      properties:
        name:
          type: string
        description:
          type: string
        specialization:
          type: string
    CurriculumUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        specialization:
          type: string

    # Chat
    ChatSession:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        messageCount:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
    ChatSessionInput:
      type: object
      properties:
        title:
          type: string
    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: number
        sessionId:
          type: number
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # Summary
...
    DebugSession:
      type: object
      required: [id, language, code, status, createdAt]
      properties:
        id:
          type: number
        language:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [clean, has_errors]
        efficiencyScore:
          type: ["number", "null"]
        errors:
          type: array
          items:
            $ref: "#/components/schemas/CodeError"
        explanation:
          type: ["string", "null"]
        bestPractices:
          type: array
          items:
            type: string
        fixedCode:
          type: ["string", "null"]
        createdAt:
          type: string
    CodeError:
      type: object
      required: [line, type, message, fix]
      properties:
        line:
          type: number
        type:
          type: string
        message:
          type: string
        fix:
          type: string
    DebugInput:
      type: object
      required: [code, language]
      properties:
        code:
          type: string
        language:
          type: string

    # Quiz
...